@@ -1,6 +1,7 @@
 ﻿// Project: Aguafrommars/TheIdServer
 // Copyright (c) 2023 @Olivier Lefebvre
 using System;
+using System.Diagnostics.CodeAnalysis;
 using System.Runtime.Serialization;
 
 namespace Duende.IdentityServer.Quickstart.UI
@@ -10,6 +11,7 @@ namespace Duende.IdentityServer.Quickstart.UI
     /// </summary>
     /// <seealso cref="System.Exception" />
     [Serializable]
+    [SuppressMessage("Major Code Smell", "S3925:\"ISerializable\" should be implemented correctly", Justification = "Obsolete")]
     public class InvalidReturnUrlException : Exception
     {
         /// <summary>
@@ -24,11 +26,6 @@ namespace Duende.IdentityServer.Quickstart.UI
         }
 
         public InvalidReturnUrlException(string message, Exception innerException) : base(message, innerException)
-        {
-        }
-
-        protected InvalidReturnUrlException(SerializationInfo serializationInfo, StreamingContext streamingContext)
-            : base(serializationInfo, streamingContext)
         {
         }
     }
