@@ -1,6 +1,5 @@
 ﻿// Project: Aguafrommars/TheIdServer
 // Copyright (c) 2023 @Olivier Lefebvre
-using Aguacongas.TheIdServer.BlazorApp.Services;
 using Microsoft.AspNetCore.Components.Routing;
 using System.Reflection;
 
@@ -8,13 +7,13 @@ namespace Aguacongas.TheIdServer.BlazorApp
 {
     public partial class App
     {
-        private readonly List<Assembly> _lazyLoadedAssemblies = new List<Assembly>
+        private readonly List<Assembly> _lazyLoadedAssemblies = new()
         {
             typeof(Pages.Index).Assembly
         };
 
-        private readonly string[] _pageKindList = new []
-        {
+        private readonly string[] _pageKindList =
+        [
             "Api",
             "ApiScope",
             "Client",
@@ -29,7 +28,7 @@ namespace Aguacongas.TheIdServer.BlazorApp
             "RelyingParties",
             "RelyingParty",
             "Settings"
-        };
+        ];
 
         protected override async Task OnInitializedAsync()
         {
@@ -37,27 +36,27 @@ namespace Aguacongas.TheIdServer.BlazorApp
             await _themeService.InitAsync().ConfigureAwait(false);
         }
 
-    private Task OnNavigateAsync(NavigationContext args)
+        private Task OnNavigateAsync(NavigationContext args)
         {
             var path = args.Path.Split("/")[0];
             if (path == "protectresource")
             {
-                return LoadAssemblyAsync("Aguacongas.TheIdServer.BlazorApp.Pages.Api.dll");
+                return LoadAssemblyAsync("Aguacongas.TheIdServer.BlazorApp.Pages.Api.wasm");
             }
 
             if (path == "identityresource")
             {
-                return LoadAssemblyAsync("Aguacongas.TheIdServer.BlazorApp.Pages.Identity.dll");
+                return LoadAssemblyAsync("Aguacongas.TheIdServer.BlazorApp.Pages.Identity.wasm");
             }
 
             var pageKind = _pageKindList.FirstOrDefault(k => path == $"{k.ToLower()}s");
             if (pageKind != null)
             {
-                return LoadAssemblyAsync($"Aguacongas.TheIdServer.BlazorApp.Pages.{pageKind}s.dll");
+                return LoadAssemblyAsync($"Aguacongas.TheIdServer.BlazorApp.Pages.{pageKind}s.wasm");
             }
 
-            pageKind = _pageKindList.FirstOrDefault(k => path == k.ToLower());
-            return pageKind != null ? LoadAssemblyAsync($"Aguacongas.TheIdServer.BlazorApp.Pages.{pageKind}.dll") : Task.CompletedTask;
+            pageKind = _pageKindList.FirstOrDefault(k => path.Equals(k, StringComparison.InvariantCultureIgnoreCase));
+            return pageKind != null ? LoadAssemblyAsync($"Aguacongas.TheIdServer.BlazorApp.Pages.{pageKind}.wasm") : Task.CompletedTask;
         }
 
         private async Task LoadAssemblyAsync(string assemblyName)
