@@ -31,15 +31,15 @@ else
                 <input asp-for="RememberMe" type="hidden" />
                 <div asp-validation-summary="ModelOnly" class="text-danger"></div>
                 <div class="form-floating mb-2">
-                    <input asp-for="Input.TwoFactorCode" class="form-control" autocomplete="off" />
-                    <label asp-for="Input.TwoFactorCode" class="form-label"></label>
-                    <span asp-validation-for="Input.TwoFactorCode" class="text-danger"></span>
+                    <input asp-for="Input!.TwoFactorCode!" class="form-control" autocomplete="off" />
+                    <label asp-for="Input!.TwoFactorCode!" class="form-label"></label>
+                    <span asp-validation-for="Input!.TwoFactorCode!" class="text-danger"></span>
                 </div>
                 <div class="form-group  mb-2">
                     <div class="checkbox">
-                        <label asp-for="Input.RememberMachine">
-                            <input asp-for="Input.RememberMachine" />
-                            @Html.DisplayNameFor(m => m.Input.RememberMachine)
+                        <label asp-for="Input!.RememberMachine!">
+                            <input asp-for="Input!.RememberMachine!" />
+                            @Html.DisplayNameFor(m => m!.Input!.RememberMachine)
                         </label>
                     </div>
                 </div>
@@ -50,7 +50,7 @@ else
         </div>
     </div>
     <p>
-        @Localizer["Don't have access to your authenticator device? You can <a href=\"{0}\">log in with a recovery code</a>.", UrlHelper.Page("./LoginWithRecoveryCode", new { returnUrl = Model.ReturnUrl })]
+        @Localizer["Don't have access to your authenticator device? You can <a href=\"{0}\">log in with a recovery code</a>.", UrlHelper.Page("./LoginWithRecoveryCode", new { returnUrl = Model.ReturnUrl })!]
     </p>
 
     @section Scripts {
