@@ -41,11 +41,11 @@ namespace Aguacongas.TheIdServer.Areas.Identity.Pages.Account
         }
 
         [BindProperty]
-        public InputModel Input { get; set; }
+        public InputModel? Input { get; set; }
 
         public bool RememberMe { get; set; }
 
-        public string ReturnUrl { get; set; }
+        public string? ReturnUrl { get; set; }
 
         public bool RedirectToReturnUrl { get; set; }
 
@@ -55,13 +55,13 @@ namespace Aguacongas.TheIdServer.Areas.Identity.Pages.Account
             [StringLength(7, ErrorMessage = "The {0} must be at least {2} and at max {1} characters long.", MinimumLength = 6)]
             [DataType(DataType.Text)]
             [Display(Name = "Authenticator code")]
-            public string TwoFactorCode { get; set; }
+            public string? TwoFactorCode { get; set; }
 
             [Display(Name = "Remember this machine")]
             public bool RememberMachine { get; set; }
         }
 
-        public async Task<IActionResult> OnGetAsync(bool rememberMe, string returnUrl = null)
+        public async Task<IActionResult> OnGetAsync(bool rememberMe, string? returnUrl = null)
         {
             // Ensure the user has gone through the username & password screen first
             var user = await _signInManager.GetTwoFactorAuthenticationUserAsync();
@@ -77,7 +77,7 @@ namespace Aguacongas.TheIdServer.Areas.Identity.Pages.Account
             return Page();
         }
 
-        public async Task<IActionResult> OnPostAsync(string userName, bool rememberMe, string returnUrl = null)
+        public async Task<IActionResult> OnPostAsync(string userName, bool rememberMe, string? returnUrl = null)
         {
             if (!ModelState.IsValid)
             {
@@ -92,7 +92,7 @@ namespace Aguacongas.TheIdServer.Areas.Identity.Pages.Account
                 throw new InvalidOperationException($"Unable to load two-factor authentication user.");
             }
 
-            var authenticatorCode = Input.TwoFactorCode.Replace(" ", string.Empty).Replace("-", string.Empty);
+            var authenticatorCode = Input!.TwoFactorCode!.Replace(" ", string.Empty).Replace("-", string.Empty);
 
             var result = await _signInManager.TwoFactorAuthenticatorSignInAsync(authenticatorCode, rememberMe, Input.RememberMachine);
 
