@@ -14,7 +14,7 @@ Copyright (c) 2023 @Olivier Lefebvre
 
 <h4>@ViewData["Title"]</h4>
 <h5>@Localizer["Below is the list of backchannel login requests awaiting your approbation."]</h5>
-@if (Model.Logins.Any() == true)
+@if (Model.Logins?.Any() == true)
 {
     <table class="table table-bordered table-striped table-sm">
         <caption class="visually-hidden">@Localizer["Pending Backchannel Login Requests"]</caption>
