@@ -21,7 +21,7 @@ Copyright (c) 2023 @Olivier Lefebvre
 </div>
 <div class="grants">
 
-    @if (Model.Grants.Any() == false)
+    @if (Model.Grants?.Any() == false)
     {
         <div class="row">
             <div class="col-sm-8">
@@ -33,7 +33,7 @@ Copyright (c) 2023 @Olivier Lefebvre
     }
     else
     {
-        foreach (var grant in Model.Grants)
+        foreach (var grant in Model.Grants!)
         {
             <div class="row grant">
                 <div class="col-2">
@@ -57,7 +57,7 @@ Copyright (c) 2023 @Olivier Lefebvre
                             <span class="expires">@Localizer["Expires:"]</span> @grant.Expires.Value.ToString("yyyy-MM-dd")
                         </div>
                     }
-                    @if (grant.IdentityGrantNames.Any())
+                    @if (grant.IdentityGrantNames?.Any() == true)
                     {
                         <div>
                             <div class="granttype">@Localizer["Identity Grants"]</div>
@@ -69,7 +69,7 @@ Copyright (c) 2023 @Olivier Lefebvre
                             </ul>
                         </div>
                     }
-                    @if (grant.ApiGrantNames.Any())
+                    @if (grant.ApiGrantNames?.Any() == true)
                     {
                         <div>
                             <div class="granttype">@Localizer["API Grants"]</div>
