@@ -1,23 +1,16 @@
 // Project: Aguafrommars/TheIdServer
 // Copyright (c) 2023 @Olivier Lefebvre
 using Aguacongas.TheIdServer.Models;
-using IdentityModel;
 using Duende.IdentityServer;
 using Duende.IdentityServer.Events;
 using Duende.IdentityServer.Services;
-using Duende.IdentityServer.Stores;
+using IdentityModel;
 using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.Extensions.Logging;
-using System;
-using System.Collections.Generic;
 using System.IdentityModel.Tokens.Jwt;
-using System.Linq;
 using System.Security.Claims;
-using System.Threading.Tasks;
 
 namespace Aguacongas.TheIdServer.UI
 {
@@ -101,7 +94,7 @@ namespace Aguacongas.TheIdServer.UI
 
             if (_logger.IsEnabled(LogLevel.Debug))
             {
-                var externalClaims = result.Principal.Claims.Select(c => $"{c.Type}: {c.Value}");
+                var externalClaims = result.Principal!.Claims.Select(c => $"{c.Type}: {c.Value}");
                 _logger.LogDebug("External claims: {@claims}", externalClaims);
             }
 
@@ -112,7 +105,7 @@ namespace Aguacongas.TheIdServer.UI
                 // this might be where you might initiate a custom workflow for user registration
                 // in this sample we don't show how that would be done, as our sample implementation
                 // simply auto-provisions new external user
-                user = await AutoProvisionUserAsync(provider, providerUserId, claims).ConfigureAwait(false);
+                user = await AutoProvisionUserAsync(provider!, providerUserId, claims).ConfigureAwait(false);
             }
 
             // this allows us to collect any additonal claims or properties
@@ -141,7 +134,7 @@ namespace Aguacongas.TheIdServer.UI
             await HttpContext.SignOutAsync(IdentityConstants.ExternalScheme);
 
             // retrieve return URL
-            var returnUrl = result.Properties.Items["returnUrl"] ?? "~/";
+            var returnUrl = result.Properties!.Items["returnUrl"] ?? "~/";
 
             // check if external login is in the context of an OIDC request
             var context = await _interaction.GetAuthorizationContextAsync(returnUrl);
@@ -172,28 +165,28 @@ namespace Aguacongas.TheIdServer.UI
             };
 
             var id = new ClaimsIdentity(provider);
-            var name = user.Identity.Name ??
+            var name = user.Identity!.Name ??
                 user.FindFirst(JwtClaimTypes.Name)?.Value ??
                 user.FindFirst(ClaimTypes.Name)?.Value ??
                 user.FindFirst(JwtSecurityTokenHandler.DefaultOutboundClaimTypeMap[ClaimTypes.Name])?.Value ??
                 user.FindFirst(ClaimTypes.NameIdentifier)?.Value ??
                 user.FindFirst(JwtSecurityTokenHandler.DefaultOutboundClaimTypeMap[ClaimTypes.NameIdentifier])?.Value;
 
-            id.AddClaim(new Claim(JwtClaimTypes.Subject, name));
-            id.AddClaim(new Claim(JwtClaimTypes.Name, name));
+            id.AddClaim(new Claim(JwtClaimTypes.Subject, name!));
+            id.AddClaim(new Claim(JwtClaimTypes.Name, name!));
 
             await HttpContext.SignInAsync(
                 IdentityConstants.ExternalScheme,
                 new ClaimsPrincipal(id),
                 props).ConfigureAwait(false);
 
-            return Redirect(props.RedirectUri);
+            return Redirect(props.RedirectUri!);
         }
 
-        private async Task<(ApplicationUser user, string provider, string providerUserId, IEnumerable<Claim> claims)>
+        private async Task<(ApplicationUser? user, string? provider, string providerUserId, IEnumerable<Claim> claims)>
             FindUserFromExternalProviderAsync(AuthenticateResult result)
         {
-            var externalUser = result.Principal;
+            var externalUser = result.Principal!;
 
             // try to determine the unique id of the external user (issued by the provider)
             // the most common claim type for that are the sub claim and the NameIdentifier
@@ -207,11 +200,11 @@ namespace Aguacongas.TheIdServer.UI
             var claims = externalUser.Claims.ToList();
             claims.Remove(userIdClaim);
 
-            var provider = result.Properties.Items["scheme"];
+            var provider = result.Properties?.Items["scheme"];
             var providerUserId = userIdClaim.Value;
 
             // find external user
-            var user = await _userManager.FindByLoginAsync(provider, providerUserId);
+            var user = await _userManager.FindByLoginAsync(provider!, providerUserId);
 
             return (user, provider, providerUserId, claims);
         }
@@ -263,7 +256,7 @@ namespace Aguacongas.TheIdServer.UI
             var identityResult = await _userManager.CreateAsync(user);
             if (!identityResult.Succeeded) throw new InvalidOperationException(identityResult.Errors.First().Description);
 
-            if (filtered.Any())
+            if (filtered.Count > 0)
             {
                 identityResult = await _userManager.AddClaimsAsync(user, filtered);
                 if (!identityResult.Succeeded) throw new InvalidOperationException(identityResult.Errors.First().Description);
@@ -280,14 +273,14 @@ namespace Aguacongas.TheIdServer.UI
         {
             // if the external system sent a session id claim, copy it over
             // so we can use it for single sign-out
-            var sid = externalResult.Principal.Claims.FirstOrDefault(x => x.Type == JwtClaimTypes.SessionId);
+            var sid = externalResult.Principal!.Claims.FirstOrDefault(x => x.Type == JwtClaimTypes.SessionId);
             if (sid != null)
             {
                 localClaims.Add(new Claim(JwtClaimTypes.SessionId, sid.Value));
             }
 
             // if the external provider issued an id_token, we'll keep it for signout
-            var id_token = externalResult.Properties.GetTokenValue("id_token");
+            var id_token = externalResult.Properties!.GetTokenValue("id_token");
             if (id_token != null)
             {
                 localSignInProps.StoreTokens(new[] { new AuthenticationToken { Name = "id_token", Value = id_token } });
