@@ -4,7 +4,7 @@ namespace Aguacongas.TheIdServer.UI
 {
     public class ExternalProvider
     {
-        public string DisplayName { get; set; }
-        public string AuthenticationScheme { get; set; }
+        public string? DisplayName { get; set; }
+        public string? AuthenticationScheme { get; set; }
     }
 }
