@@ -7,10 +7,10 @@ namespace Aguacongas.TheIdServer.UI
     public class LoginInputModel
     {
         [Required]
-        public string Username { get; set; }
+        public string? Username { get; set; }
         [Required]
-        public string Password { get; set; }
+        public string? Password { get; set; }
         public bool RememberLogin { get; set; }
-        public string ReturnUrl { get; set; }
+        public string? ReturnUrl { get; set; }
     }
 }
