@@ -4,6 +4,6 @@ namespace Aguacongas.TheIdServer.UI
 {
     public class LogoutInputModel
     {
-        public string LogoutId { get; set; }
+        public string? LogoutId { get; set; }
     }
 }
