@@ -4,6 +4,6 @@ namespace Aguacongas.TheIdServer.UI
 {
     public class RedirectViewModel
     {
-        public string RedirectUrl { get; set; }
+        public string? RedirectUrl { get; set; }
     }
 }
