@@ -4,10 +4,10 @@ namespace Aguacongas.TheIdServer.Duende.Quickstart.Ciba
 {
     public class InputModel
     {
-        public string Button { get; set; }
-        public IEnumerable<string> ScopesConsented { get; set; }
-        public string Id { get; set; }
-        public string Description { get; set; }
+        public string? Button { get; set; }
+        public IEnumerable<string>? ScopesConsented { get; set; }
+        public string? Id { get; set; }
+        public string? Description { get; set; }
 
     }
 }
