@@ -2,7 +2,7 @@
 {
     public class ResourceViewModel
     {
-        public string Name { get; set; }
-        public string DisplayName { get; set; }
+        public string? Name { get; set; }
+        public string? DisplayName { get; set; }
     }
 }
