@@ -6,10 +6,10 @@ namespace Aguacongas.TheIdServer.UI
 {
     public class ConsentInputModel
     {
-        public string Button { get; set; }
-        public IEnumerable<string> ScopesConsented { get; set; }
+        public string? Button { get; set; }
+        public IEnumerable<string>? ScopesConsented { get; set; }
         public bool RememberConsent { get; set; }
-        public string ReturnUrl { get; set; }
-        public string Description { get; set; }
+        public string? ReturnUrl { get; set; }
+        public string? Description { get; set; }
     }
 }
