@@ -6,17 +6,17 @@ namespace Aguacongas.TheIdServer.UI
 {
     public class ConsentViewModel : ConsentInputModel
     {
-        public string ClientName { get; set; }
-        public string ClientUrl { get; set; }
-        public string ClientLogoUrl { get; set; }
+        public string? ClientName { get; set; }
+        public string? ClientUrl { get; set; }
+        public string? ClientLogoUrl { get; set; }
 
-        public string PolicyUrl { get; set; }
+        public string? PolicyUrl { get; set; }
 
-        public string TosUrl { get; set; }
+        public string? TosUrl { get; set; }
 
         public bool AllowRememberConsent { get; set; }
 
-        public IEnumerable<ScopeViewModel> IdentityScopes { get; set; }
-        public IEnumerable<ScopeViewModel> ApiScopes { get; set; }
+        public IEnumerable<ScopeViewModel>? IdentityScopes { get; set; }
+        public IEnumerable<ScopeViewModel>? ApiScopes { get; set; }
     }
 }
