@@ -7,13 +7,13 @@ namespace Aguacongas.TheIdServer.UI
     public class ProcessConsentResult
     {
         public bool IsRedirect => RedirectUri != null;
-        public string RedirectUri { get; set; }
-        public Client Client { get; set; }
+        public string? RedirectUri { get; set; }
+        public Client? Client { get; set; }
 
         public bool ShowView => ViewModel != null;
-        public ConsentViewModel ViewModel { get; set; }
+        public ConsentViewModel? ViewModel { get; set; }
 
         public bool HasValidationError => ValidationError != null;
-        public string ValidationError { get; set; }
+        public string? ValidationError { get; set; }
     }
 }
