@@ -4,9 +4,9 @@ namespace Aguacongas.TheIdServer.UI
 {
     public class ScopeViewModel
     {
-        public string Value { get; set; }
-        public string DisplayName { get; set; }
-        public string Description { get; set; }
+        public string? Value { get; set; }
+        public string? DisplayName { get; set; }
+        public string? Description { get; set; }
         public bool Emphasize { get; set; }
         public bool Required { get; set; }
         public bool Checked { get; set; }
