@@ -6,6 +6,6 @@ namespace Aguacongas.IdentityServer.UI.Device
 {
     public class DeviceAuthorizationInputModel : ConsentInputModel
     {
-        public string UserCode { get; set; }
+        public string? UserCode { get; set; }
     }
 }
