@@ -6,7 +6,7 @@ namespace Aguacongas.IdentityServer.UI.Device
 {
     public class DeviceAuthorizationViewModel : ConsentViewModel
     {
-        public string UserCode { get; set; }
+        public string? UserCode { get; set; }
         public bool ConfirmUserCode { get; set; }
     }
 }
