@@ -14,8 +14,8 @@ namespace Aguacongas.TheIdServer.UI
     {
         public async Task<IActionResult> Index()
         {
-            var localAddresses = new string[] { "127.0.0.1", "::1", HttpContext.Connection.LocalIpAddress.ToString() };
-            if (!localAddresses.Contains(HttpContext.Connection.RemoteIpAddress.ToString()))
+            var localAddresses = new string[] { "127.0.0.1", "::1", HttpContext.Connection.LocalIpAddress!.ToString() };
+            if (!localAddresses.Contains(HttpContext.Connection.RemoteIpAddress!.ToString()))
             {
                 return NotFound();
             }
