@@ -2,7 +2,6 @@
 // Copyright (c) 2023 @Olivier Lefebvre
 using Duende.IdentityServer.Models;
 using Microsoft.AspNetCore.Mvc;
-using System;
 
 namespace Aguacongas.TheIdServer.UI
 {
@@ -18,10 +17,10 @@ namespace Aguacongas.TheIdServer.UI
                && !context.RedirectUri.StartsWith("http", StringComparison.Ordinal);
         }
 
-        public static IActionResult LoadingPage(this Controller controller, string viewName, string redirectUri)
+        public static IActionResult LoadingPage(this Controller controller, string? viewName, string? redirectUri)
         {
             controller.HttpContext.Response.StatusCode = 200;
-            controller.HttpContext.Response.Headers["Location"] = "";
+            controller.HttpContext.Response.Headers.Location = "";
 
             return controller.View(viewName, new RedirectViewModel { RedirectUrl = redirectUri });
         }
