@@ -14,7 +14,7 @@ Copyright (c) 2023 @Olivier Lefebvre
                 <div class="client-logo"><img src="@Model.ClientLogoUrl" alt="logo"></div>
             }
             <h1>
-                @Localizer["{0} <small>is requesting your permission</small>", @Model.ClientName]
+                @Localizer["{0} <small>is requesting your permission</small>", @Model.ClientName!]
             </h1>
         </div>
     </div>
@@ -24,7 +24,7 @@ Copyright (c) 2023 @Olivier Lefebvre
         <div class="row">
             <div class="col-sm-8">
                 <p>
-                    @Localizer["Please confirm that the authorization request quotes the code: \"{0}\".", Model.UserCode]
+                    @Localizer["Please confirm that the authorization request quotes the code: \"{0}\".", Model.UserCode!]
                 </p>
             </div>
         </div>
@@ -39,7 +39,7 @@ Copyright (c) 2023 @Olivier Lefebvre
 
                 <div>@Localizer["Uncheck the permissions you do not wish to grant."]</div>
 
-                @if (Model.IdentityScopes.Any())
+                @if (Model.IdentityScopes?.Any() == true)
                 {
                     <div class="panel panel-default consent-buttons">
                         <div class="panel-heading">
@@ -55,7 +55,7 @@ Copyright (c) 2023 @Olivier Lefebvre
                     </div>
                 }
 
-                @if (Model.ApiScopes.Any())
+                @if (Model.ApiScopes?.Any() == true)
                 {
                     <div class="panel panel-default">
                         <div class="panel-heading">
