@@ -7,8 +7,8 @@ Copyright (c) 2023 @Olivier Lefebvre
 @model ErrorViewModel
 
 @{
-    var error = Model?.Error?.Error != null ? Localizer[Model?.Error?.Error] : null;
-    var errorDescription = Model?.Error?.ErrorDescription != null ? Localizer[Model?.Error?.ErrorDescription] : null;
+    var error = Model?.Error?.Error != null ? Localizer[Model.Error.Error] : null;
+    var errorDescription = Model?.Error?.ErrorDescription != null ? Localizer[Model.Error.ErrorDescription] : null;
     var request_id = Model?.Error?.RequestId;
 }
 
