@@ -14,7 +14,7 @@ Copyright (c) 2023 @Olivier Lefebvre
     <form asp-area="Identity" asp-page="/Account/Logout" asp-route-returnUrl="@Url.Page("/Index", new { area = "" })" method="post" id="logoutForm" class="navbar-right">
         <ul class="nav navbar-nav navbar-right">
             <li class="nav-item">
-                <a asp-area="Identity" asp-page="/Account/Manage/Index" title="Manage">@Localizer["Hello {0}!", UserManager.GetUserName(User)]</a>
+                <a asp-area="Identity" asp-page="/Account/Manage/Index" title="Manage">@Localizer["Hello {0}!", UserManager.GetUserName(User)!]</a>
             </li>
             <li class="nav-item">
                 <button type="submit" class="btn btn-link navbar-btn navbar-link">@Localizer["Logout"]</button>
