@@ -1,7 +1,7 @@
 ﻿// Project: Aguafrommars/TheIdServer
 // Copyright (c) 2023 @Olivier Lefebvre
 using System;
-using System.Runtime.Serialization;
+using System.Diagnostics.CodeAnalysis;
 
 namespace Aguacongas.IdentityServer.Admin.Models
 {
@@ -10,6 +10,7 @@ namespace Aguacongas.IdentityServer.Admin.Models
     /// </summary>
     /// <seealso cref="System.Exception" />
     [Serializable]
+    [SuppressMessage("Major Code Smell", "S3925:\"ISerializable\" should be implemented correctly", Justification = "Obsolete")]
     public class RegistrationException : Exception
     {
         /// <summary>
@@ -27,15 +28,6 @@ namespace Aguacongas.IdentityServer.Admin.Models
         public RegistrationException(string errorCode, string message) : base(message)
         {
             ErrorCode = errorCode;
-        }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="RegistrationException"/> class.
-        /// </summary>
-        /// <param name="info">The <see cref="T:System.Runtime.Serialization.SerializationInfo" /> that holds the serialized object data about the exception being thrown.</param>
-        /// <param name="context">The <see cref="T:System.Runtime.Serialization.StreamingContext" /> that contains contextual information about the source or destination.</param>
-        protected RegistrationException(SerializationInfo info, StreamingContext context) : base(info, context)
-        {
         }
     }
 }
